crypto_common: decrypt v2 data blocks, and option for encrypting
This changes the cryptoPure block-related interfaces to pass in the
TLF crypt key and blcok server half explicitly, so that `CryptoCommon`
can make a decision about the right way to construct the block key and
encrypt/decrypt the data.  With this commit, clients are able to
understand v2 blocks.

This also adds an option to `CryptoCommon` for encrypting the data in
the new format, though it's off by default except in select tests.  A
future commit will wire the option to a command line argument.

Issue: KBFS-3458